added tests for calculate lfc3d by screen
