+++ tests/variables.py
@@ -45,3 +45,20 @@ all_mouse_uniprots = [
     'D3YYC3', 'Q9Z148', 'Q3UL97', 'Q06180', 
     'Q62318', 'O54864', 
 ]
+
+from sklearn.feature_extraction.text import TfidfVectorizer
+from sklearn.metrics.pairwise import cosine_similarity
+
+# use a fuzzy compare function because there are some minor discrepancies with capitalization and ordering
+# could also use difflib.SequenceMatcher, Levenshtein.ratio, ngrams word_tokenize (Jaccard similarity)
+# but the TfidfVectorizer cosine_similarity is the fastest for large documents
+def fuzzy_compare(file1, file2):
+    with open(file1, 'r') as f1, open(file2, 'r') as f2:
+        text1 = f1.read()
+        text2 = f2.read()
+
+    vectorizer = TfidfVectorizer().fit_transform([text1, text2])
+    vectors = vectorizer.toarray()
+    similarity = cosine_similarity(vectors)
+    score = similarity[0][1]
+    return score
